added index file for component with export
