+++ src/App.js
@@ -1,22 +1,19 @@
 import React from 'react'
 
-import Header from './components/Header';
-import Map from './pages/Map';
-import Profile from './pages/Profile';
-
-// import logoImg from './assets/images/logo-img.svg';
-// import logoText from './assets/images/logo-text.svg';
-// import LoginForm from './components/LoginForm';
-// import RegistrationForm from './components/RegistrationForm';
+import { Header } from './components/Header';
+import { Home } from './pages/Home';
+import { Map } from './pages/Map';
+import { Profile } from './pages/Profile';
 
 const PAGES = {
+  home: <Home />,
   map: <Map />,
   profile: <Profile />
 }
 
 class App extends React.Component {
 
-  state = { currentPage: 'map' };
+  state = { currentPage: 'home' };
 
   changePage = (page) => {
     this.setState({ currentPage: page });
@@ -32,20 +29,6 @@ class App extends React.Component {
           </section>
         </main>
       </>
-      // <div className='wrapper'>
-      //   <div className='container app-section'>
-      //     <div className='mainPage_logoBlock'>
-      //       <div className='logo-block'>
-      //         <img src={logoImg} alt='loft taxi logo' className='logo-block__img' />
-      //         <img src={logoText} alt='loft taxi logo' className='logo-block__text' />
-      //       </div>
-      //     </div>
-      //     <div className='mainPage_formBlock'>
-      //       <LoginForm />
-      //       <RegistrationForm />
-      //     </div>
-      //   </div>
-      // </div>
     );
   }
 }
