+++ src/pages/Map/Map.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export default function Map() {
+export function Map() {
   return (
     <h1>Map page</h1>
   )
