+++ src/pages/Map/index.js
@@ -1,0 +1,1 @@
+export { Map } from './Map';
