+++ src/pages/Profile/Profile.spec.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Profile from './'
+import { Profile } from './'
 import { render } from '@testing-library/react'
 
 describe("Profile", () => {
